Add partial support for looking up attributes on Type::Intersect
Summary:
Adds support for attribute lookup on Type::Intersect in the simple case in which exactly one type in an intersection has the attribute. Otherwise, we use the intersection's fallback type, preserving previous behavior.

This change on its own is a no-op, because we're not yet creating any intersection types.

Reviewed By: grievejia

Differential Revision: D86880947

fbshipit-source-id: fdd363238f6ebdfdf9aca11457fba7cfac1839fe