Create Type::Intersect in narrowing operations that intersect types with a non-Never fallback
Summary:
Previously, if we needed to create a non-Never intersection of two types for a type narrowing operation, we approximated this by falling back to one of the types. This meant that we effectively cast the result to the fallback type, throwing away the other type. Now we instead produce an intersection of the two types. The intersection stores the previous fallback type, and we fall back to it in cases where I'm not sure how to handle intersections and/or we need further discussion on what behavior we want.

This fixes a bunch of type narrowing bugs in two real-world projects (helion and castfit), showing that even this partial support for intersections is enough to be useful.

Fixes https://github.com/facebook/pyrefly/issues/526.

Fixes https://github.com/facebook/pyrefly/issues/814.

Relevant issues that this diff doesn't fully fix:
* https://github.com/facebook/pyrefly/issues/235
* https://github.com/facebook/pyrefly/issues/1140

Reviewed By: grievejia

Differential Revision: D86880951

fbshipit-source-id: 9fc7b76ad75cd4666fa88b775d79912f4ab865b4