Add a fallback type to Type::Intersect
Summary: I want to add some basic support for intersection types in order to fix a bug that is generating a bunch of false positives in pytorch/helion. This diff adds a fallback type to Type::Intersect to make it possible to add partial support for intersections while falling back to the non-intersection type we would've used previously in places where we're not yet prepared to handle intersections.

Reviewed By: grievejia

Differential Revision: D86880952

fbshipit-source-id: cc078af650418a1cc385d58ce5a915300ead6526