Add a test case of return type inference + functools.wraps
Summary:
This tracks a bug report in issue 1448. This is a bit of a weird case: because it
involves return type inference (and pretty much has to, the return type `_Wrapped` is
private so people *shouldn't* use it directly, although they certainly could), the
only existing type checker to reference is Pyright.

It looks like Pyright actually hardcodes this in a way that I don't think we can
easily do in Pyrefly; it seems to make the entire decorator just preserve signatures
when it sees `_Wrapped` in a decorator return type. I don't think we can take exactly
the same approach very easily right now, but I do think we can do something related
to eliminate the FP, see the next diff.

Reviewed By: rchen152

Differential Revision: D87031255

fbshipit-source-id: 0a81b0d4cfecfb73446e4cdab236afe9cf5b4e82