Eliminate misleading `apply_decorator` funciton
Summary:
I want to patch function decorator handling - which is a mostly type-level
concern - with logic that I can't implement unless I introduce redundant
clones in the current code because we pass `decoratee` as a `Type` which
consumes it, and that prevents me from implementing logic that looks at both
the decorator application result and the decoratee at the same time.

This came up because I want to improve handling of decorators that return
`functools._Wrapper`, which requires applying the decorator and only then
deciding what to do.

But this is a bit silly - the only reason we are passing an owned `decoratee`
to apply_decorator is so that we can return it early in the class case. But
we *always* skip the application in the class case, which means there's no
point even having an `apply_decorator`, the only thing this accomplishes is
that it puts the TODO in a worse place - it's much better if the TODO to
think about class decorator application lives at the class binding solve than
buried in a helper.

Reviewed By: rchen152

Differential Revision: D87031257

fbshipit-source-id: e90c70e5af1ef0cdd0fb12cd1e314c4e8d6fc592