Implement a simple fix for issue 1448
Summary:
I decided it is neither wise nor easy to implement a quick fix for Pyrefly
to exactly replicate the behavior of Pyright on `functools.wraps`; Pyright
treats the entire decorator as signature-preserving, but that's actually
wrong in cases where the inner function mangles arguments which feels
risky, plus it's a little tricky to implement.

The approach here is to treat `functools.wraps` *itself* as a signature
preserving decorator, which means that the decorator *using* it winds up
preserving the signature of the inner function.

This means that an untyped decorator library might lose some type information,
but it also avoids the unsoundness in Pyright's handling becaue if the inner
function mangles the signature (including using well-annotated code to do so)
we'll track the mangling correctly.

I'm not 100% sure of this approach, and we could potentially converge with
Pyright in the future, but I'm confident that this solution is a strict improvement
since `functools.wraps` is in fact known to preserve the signature whereas the
Pyright solution is arguably wrong in many cases.

(Note: the pytorch delta is a 1-for-1 change on a per-error basis - every error removed
is replaced by an equivalent but more readable error no longer involving `_Wrapped`,
it's effectively a no-op other than maybe a minor UX improvement)

Reviewed By: rchen152

Differential Revision: D87031254

fbshipit-source-id: 41348da93bf55b01c6c825434da1c992437d5a4f