index overridden methods
Summary:
Go to implementations is a LSP feature that finds reimplementations of a method. see [this post](https://fb.workplace.com/groups/657557118131740/permalink/1744377449449696/) for more details. this stack will attempt to implement it.

the general approach will be
- index overridden methods
- look for these in find-references (so rename works on them)
- implement go-to implementation reusing the same logic

this diff adds the index. I did a lot of testing on a large instagram file (see test plan). from our findings, we filter out a lot of things. the index ends up a lot smaller than it started at, with 284928 keys (classes) and 400897 methods (overrides).

Since this is a SmallMap keyed on TextRange, with value of Vec<(TextRange, ModulePath)>

TextRange is 8 bytes, ModulePath is 16 bytes

285000 * 8 bytes = 2.28MB
400000 * 24 bytes = 9.6MB

This seems like a rounding error in checking a codebase of this size.

other ideas:
- the index could store other things like class hierarchy (D86892234). this would help us when we implement type hierarchy. but this index is likely far too large for every object.
- we could not index at all, and find transitive classes on-the-fly during find-references or go-to implementations. this seems much harder.

Reviewed By: jvansch1

Differential Revision: D86926526

fbshipit-source-id: e1019175658cc617b98486d951c354bc744ef384