Make sure narrow keys have unique usage ranges
Summary:
Changes the ranges of some narrow keys to avoid duplicate usage ranges. Note that we previously got away with some duplicates because most narrows had a unique test expression and therefore a unique test range. I want to start recognizing test expressions that are saved to a local variable, and if the same variable is used multiple times, the test range is the same.

For https://github.com/facebook/pyrefly/issues/1537.

Reviewed By: stroxler

Differential Revision: D86926291

fbshipit-source-id: bf2fa8571e9546a9e6dd48585762aabb6da198b9