Refactor call_infer to accept call_range parameter
Summary:
we need to access the range of the function call in order to index calls to constructors.

the place we want to index these is in create_class, where we currently only have access to the arguments range (what range is). this diff makes a helper function so we can avoid changing all callsites of call_infer but still pass the function args when we need to.

Reviewed By: stroxler

Differential Revision: D87100848

fbshipit-source-id: 4cbfa7d43d381b11c3295b7ec24f806086c93125