Add support for __init__ constructor references
Summary: This change enables tracking references to __init__ methods when classes are instantiated. When a class is called to create an instance and it has a custom __init__ method, the constructor call site now records a reference to that __init__ method. This allows LSP features like "find all references" to show where a class's __init__ is being invoked through constructor calls.

Reviewed By: stroxler

Differential Revision: D87101155

fbshipit-source-id: 5892ab6bb97e1cfd3e77d97f7cc6e2d8db749581