Merge AttributeBase1::TypeQuantified and AttributeBase1::ClassObject(ClassBase::Quantified)
Summary: I noticed while working on the previous diff in this stack that these two are very similar. Merging them simplifies the code a bit, while also making it less likely that we'll accidentally treat type[SomeTypeVar] differently from type[SomeClass], which was the root cause of the bug the previous diff fixed.

Reviewed By: stroxler

Differential Revision: D87154052

fbshipit-source-id: f3291d5e66d0fb4eca313406285c8569f5b2ca66