Use functional logic for `is_inherited`
Summary:
Return the flag from the block rather than mutating a function-scope variable.

I don't think the code is actually more pleasing to the eye this way, it's not
very readable in the before *or* after state, but functional logic tends to be
easier to rework because the types say what it does whereas mutations live
outside the type system so they are much easier to mess up in a refactor without
the compiler warning you.

In addition, this way `is_inherited` doesn't have to be mutable for the entire
body of `calculate_class_field`, which is hundreds of lines long.

Reviewed By: rchen152

Differential Revision: D87284706

fbshipit-source-id: eae0225a15d76f644bc85ca62e965a1b9b44a103