Split `get_class_member` into two interfaces
Summary:
Most callsites were just getting the value; I think `get_class_member`
should just return the value and we should have a dedicated helper for
the relative handful of cases (all of them weird cases with very
special semantics like `__new__` and metaclass lookups) where the
defining class matters.

With this change, at the cost of one simple helper a bunch of callsites
get a bit easier to read, and we can make it explicit that the defining
class portion of the logic is private to `class_field.rs`, not part of
the public interface.

Reviewed By: rchen152

Differential Revision: D87284708

fbshipit-source-id: 8a6d9e610f4e699d718f8ceb99bdca348bb3e464