Make it much clearer what typed dict field semantics are
Summary:
This diff eliminates a tiny preexisting check of an initialized class field
in a typed dict in favor of a more comprehensive test of the behavior.

Our current behavior today is to complain on any field that isn't a bare
annotation - this is consistent with other type checkers, which is good,
although the error messages aren't great.

It's also a bit funny how we handle downstream
attempts to read `g`, `h`, and `cv`; we aren't very
consistent with ourselves or other type checkers which just treat these
as Unkown. I'm planning to simplify quite a bit in a refactor where we separate
typed dict field calculation from the more normal path.

Reviewed By: rchen152

Differential Revision: D87297638

fbshipit-source-id: 771b221296647847d693441fe1bc7bb363d78231