Explicitly mark `class_field` with a validation-only section
Summary:
The `calculate_class_field` funciton is enormous, and I can't improve it much
in one diff so I need to chip away at the complexity.

One thing that will help a lot is to group the logic into the critical bits that
affect later analysis (every such line of code is orders of magnitude more complex
to reason about because it can potentially interact with a line of code anywhere
in Pyrefly) versus validation hooks that can be fully understood by just reading
it because it never interacts with downstream.

Added a comment to clearly note that once the `class_field` object - which is
not mutable - is created, everything below that point is just validation.

has a ret

Reviewed By: rchen152

Differential Revision: D87284709

fbshipit-source-id: aff183dd883b7da78e4563fb28029d66a58dad95