Add new send_notification test API
Summary:
The new API uses Rust types to ensure the JSON is valid and send the appropriate "method" automatically.

This makes it a bit safer to write these tests, since passing the wrong JSON value will be caught at runtime. It also removes a bit of boilerplate, making the test code itself a bit shorter.

Reviewed By: grievejia

Differential Revision: D87277235

fbshipit-source-id: 0d37e91134a248fd44f8ed83fa3d8704ddc3fb47