Improve expect_request / expect_response APIs
Summary:
Similar to the send_request and send_response APIs, we can use types to make things a bit better.

I also split expect_response_error into a separate function. In almost every case, we expect the response to have a result. This removes a fair bit of boilerplate.

Updating all the callers is a lot of code, but it's all mechanical changes.

Reviewed By: stroxler

Differential Revision: D87301143

fbshipit-source-id: 5e519c476e46184a4c3eedd9f1f2ca18a58e25e9