Split out typed dict field logic
Summary:
Today I started chipping away with some tiny refactors of
`calculate_class_field`; my ultimate goal is to make not only the
data but the resulting datatypes much more aware of semantics (the
ClassFieldDefinition enum has a lot of easily encoded information that
gets torn up into a bunch of fields and it's very hard to understand).

As I was doing this, I realized that typed dict field handling is in
the way, and I checked other type checkers: all of them completely ban
typed dicts that contain non-typed-dict-field fields (e.g. classmethods,
classvars, and static methods).

I think we should do the same, and if we do then we can make the typed
dict handling special-cased and much simpler, in the process removing that
as a huge edge case in the rest of the class field logic.

This diff begins working on that; unfortunately because the existing logic
is coupled to the existing variable definitions, I have to make things
worse before I make them better - as of this diff, I've largely copied the
existing `calculate_class_field` code although I was able to strip out many
blocks that are obviously irrelevant to typed dicts. I'll spend the rest of
the refactor cleaning this up.

Even though this makes a mess of the typed dict code by duplicating everything,
it immediately gives us pretty sizeable readability wins in the main
`calculate_class_field` function, because we can delete all the typed-dict-specific
logic which is great. A couple of error messages also improve since we are
no longer coupling typed dict and named tuple logic.

Reviewed By: rchen152

Differential Revision: D87307713

fbshipit-source-id: 973c28c0c167f54a9e487a7e182f3e347652f805