Fix incorrect gotodef destination on implicitly imported names
Summary:
When seeing a bare `import `statement such as `import x.y`, Pyrefly creates a `Binding::Module(...)` for this import. The binding carries 3 pieces of data and only the first 2 are relevant to us. The interpretation of the data is actually non-straightforward:
- If we see `Binding::Module(x.y, x.y)`, the module being bound would be `x.y`.
- If we see `Binding::Module(x.y, x)`, the module being bound is actually just `x`.

The solver code [knows these rules](https://www.internalfb.com/code/fbsource/[f965b1a32c2d]/fbcode/pyrefly/pyrefly/lib/alt/solve.rs?lines=3353-3361) when solving, but the gotodef logic doesn't currently implement the same rule, and instead naively assumes that the first piece of data `x.y` is always the module being bound, which leads to incorrect gotodef results when resolving the name reference `x`.

NOTE: The module binding setup is indeed confusing -- it's actually not hard to end up with implementing the wrong interpretation.

This diff fixes the issue by changing the gotodef logic to interpret `Binding::Module` in the same way as the solver logic.

Reviewed By: ndmitchell, kinto0

Differential Revision: D87277053

fbshipit-source-id: 19a53bfd988fa61b43ec13261ccf0cacfd5c5681