Simplifiy typed dict read-only-reason
Summary:
Most of the logic in the `ReadOnlyReason` helper is not
relevant to typed dict fields; it's simpler to just directly
check for the `ReadOnly` qualifier, which is the only valid
way for a typed dict field to be read-only.

Doing so eliminates the need for quite so many junk variables,
which are introduced by the attempt to share logic with the much
more complex general case; as a result we can remove the `class`
argument which is now unused, and it unblocks simplifying the
`value_ty` away.

Reviewed By: yangdanny97

Differential Revision: D87307706

fbshipit-source-id: c79f6dee7349e6b4250bc1a5cabc7fb639dfae8d