Simplify out the `value_ty` in typed dict code
Summary:
It was always just `Any` anyway - we don't need it.

I added a TODO because it's possible there might be cases where a
bare qualifier annotation leads to silent implicit Any; this is very
low priority though so I'm probably not going to try to figure it
out right now.

Reviewed By: yangdanny97

Differential Revision: D87307708

fbshipit-source-id: 414069bf16f355c8c031b5613d4b58014a0c824e