Fix a panic in typed dict logic
Summary:
We incorrectly assumed that if we synthesize an `__init__`, it's safe to assume
`get_dunder_init_helper` always finds it.

That's not the case - if the user defined their own `__init__` (which is disallowed
but can occur) then the lookup can fail.

Since this is very edge-case-y, for now I just put in a TODO and use `Any` - if
the user actually tried to do this it's a type error so I think it's fine to
just give up for now, although we could do better.

Reviewed By: yangdanny97

Differential Revision: D87362443

fbshipit-source-id: 6d00d2b6d3496e3cd4b9c9f1f00518169945d7b3