Simplify typed dict init fetching
Summary:
If we're going to unwrap, it's better to do it in the class field code than in
the `call` code. It turns out this `unwrap` is not safe - we keep adding layers
of logic, and it can fail for reasons the authors didn't forsee, having a synthesized
field isn't enough.

Will address that in the next diff.

Reviewed By: yangdanny97

Differential Revision: D87362442

fbshipit-source-id: 3f41ced15437b421c8268739c4ee843cca8ccbe1