fix literal string protocol subset
Summary:
The signature of typing.Container is weird, if we tried to match it strictly we would have issues due to the input type of `Container.__contains__` being `object` whereas the input type of `str.__contains__` being `str`.

`str` matches `Container` by direct inheritance, but we can't convert `LiteralString` to `str`. since we want to be able to unwrap a `LiteralString` into a `Iterable[LiteralString]`, and converting to `str` would produce an `Iterable[str]` instead.

fixes https://github.com/facebook/pyrefly/issues/1598

Reviewed By: rchen152

Differential Revision: D87371859

fbshipit-source-id: 7bec3c003b089ffbf707e3383f0bd6f2ef269caf