Register artificial callees for decorators
Summary:
To handle decorators, pysa needs to know the callee for each decorator. Those are associated with artificial expression using the "for-decorated-target" expression identifier.

This diff generates those extra callees for decorators on the pyrefly side, and parses those in pysa so they can be used in the higher order call graph building.

Reviewed By: tianhan0

Differential Revision: D87344249

fbshipit-source-id: 9a007fce95b7911cfbe389ad67b401eea2845b1d