Use assignment statement locations for `__setitem__`
Summary:
Currently we just use the subscript expression location, but we need to stay
consistent with the call graphs built in `callGraphBuilder.ml`.

An alternative solution is to use the subscript expression location in
`callGraphBuilder.ml`, which is easier but I think the "correct" location is
indeed the statement locations, since we transform `x[a] = b` into
`x.__setitem__(a, b)`.

Reviewed By: arthaud

Differential Revision: D87373885

fbshipit-source-id: c23f915c48e721ebb0c08a5424ccceee1b5315b7