Set up TypeOutput trait.
Summary:
Creates a TypeOutput trait that will be used to implement different output types in a later diff. The idea here is that we want to be able to use the logic in fmt_helper for various use cases.

Until now, the logic in fmt_helper has been hardcoded to just write to a formatter which as been sufficient. However, we now have another need for this logic. We also want to be able to take a type and break it down into clickable parts along with their location to enable goto_type_definition on hints.

The easiest way we have thought of to do this is to create this output trait. For now we will work on replacing and existing logic in fmt_helper with an implementation of this. Once this is completed, we will then use it to make a new implementation which will collect the types and their locations.

Reviewed By: ndmitchell, kinto0

Differential Revision: D87451183

fbshipit-source-id: 365a2ef8bed0320f74c4516cf4c815a3e362995b