implement fmt_helper_generic
Summary:
This diff makes no functional change and is just replacing the existing logic of fmt_helper with DisplayOutput.

This diff is the actual implementation of DisplayOutput within the fmt_helper section. We create a new fmt_helper_generic function which is designed to be used for various purposes and take in a generic output. The existing fmt_helper will call this function and pass in a DisplayOutput struct rather than the formatter.

There are several other files that rely on this formatter which need to be updated as well.

Reviewed By: ndmitchell

Differential Revision: D87451215

fbshipit-source-id: 886d5115d15ee56dd403fabde96dd15b827ea94f