Fix a contextual typing bug in class fields
Summary:
We should only bypass contextual hints from the parent class
when an annotated class field actually has a type in the annotation,
which may not be the case if it is a qualifier-only annotation.

Not only that, if there is in fact an annotated type we shouldn't
type check the assigned expression at all, because the annotation
always takes precedence anyway.

Note that there are separate bindings (from the normal ast traversal
of the method / class body this came from) that will verify the
assignment is legal, so removing it from class_field is not going
to cause false negatives.

----------

Note the conformance change - I'm pretty sure this is a case where
we were checking the same binding twice (which happens regularly in
the class field logic, which duplicates bindings), but because of the bug
fixed here one of the checks used a different hint from the other and the
errors were not deduped. The new output seems better.

Reviewed By: grievejia

Differential Revision: D87448819

fbshipit-source-id: 7bf0fe72b475b8fe337b96122860ab90bce10c7d