Simplify the second match in `calculate_class_field`
Summary:
It was weird that we had a double-match, where the first two cases of the
initial match were guaranteed to produce `None` values that predetermined the
second match; I think the root cause was that we hadn't extracted a nice helper
like `attribute_expr_infer`, which meant the double match was the only way to
avoid code duplication.

Now that we have a helper, it's a lot simpler to just define the actual result
instead of going through two passes and having to cross-reference them; only the
case of non-private, non-annotated attributes requires looking at inheritance
and that's now much clearer from the match structure.

Reviewed By: grievejia

Differential Revision: D87448822

fbshipit-source-id: c59cf9e457c2090394f20660d90c633fe432c387