Make exect_publish_diagnostic_uri take a Url
Summary:
I noticed it was taking a str, when the other method expecting a url takes a Url object, so I changed this for consistency.

I tweaked the code to get the cell_uri (which also returned a str) to return a Url object as well. (In the process, I learned what the original author must have learned before, which is that the url library is a bit prickly!)

Also drive-by renamed the method from "exact_uri" to just "uri" -- in the LSP all the URIs are exact. The new name mirrors did_open/did_open_uri

Reviewed By: kinto0

Differential Revision: D87375559

fbshipit-source-id: ab35ce31a39226c5a5edd6a952267fd96289671f