Merge TestServer into TestClient
Summary:
The noise in this change is renaming interaction.server to interaction.client everywhere.

The signal in this change is in object_model.rs, where I have eliminated TestServer and rolled it into TestClient.

I think I understand why these were originally split up, but I think it's much clearer to have (1) the "server" thread, which is running the actual LSP server and (2) the "client" which is communicating with the server over a pair of channels.

In practice, the TestClient and TestServer were sharing too much anyway -- they both needed the same root, they both needed the same request_idx.

It's a bit nicer to not have to split up the Connection object. This change also makes some follow-up improvements possible. In particular, I am interested in supporting an API where sending a request returns a "request handle" object, which allows us to chain the "expect response" call, propagating type information.

Reviewed By: ndmitchell

Differential Revision: D87511845

fbshipit-source-id: 75170a87b96907ab3f2cd73d6534704fbe597c05