Consolidate annotation handling and promotion
Summary:
These two blocks already duplicated (in a not very readable way) the
same check against an annotation that contains a type.

Doing them together makes it all clearer and reduces the number of
transformations you have to wade through to track what's going on.

Now we don't need a comment telling us there's an optimization for
`annotation`, because it's no longer an optimization we're just
doing it directly.

Reviewed By: grievejia

Differential Revision: D87559428

fbshipit-source-id: 3ad55f1fe53ff1357afa40228e1f63d98f2b01a3