Extract `validate_direct_annotation`
Summary:
This code is fairly modular, it doesn't mutate anything
(other than `errors`) so there are no tricky semantics or interactions
with downstream to understand and there's no reason to inline it
and clutter up `calculate_class_field`.

Note that these validations can't be moved to the bottom of the
function without a small perf hit, since we consume `direct_annotation`
to combine it with `inherited_annotation` later and we do so with
no clone - as a result the validations have to happen in the middle.

Reviewed By: grievejia

Differential Revision: D87559433

fbshipit-source-id: 12321d2ec3b8e3092af942dac620cc3386719ba3