Extract `direct_annotation` from the main field_definition match
Summary:
I'm trying to chip away at the complexity of `calculate_class_field` until
it becomes possible to change the representation we use for class fields to
an enum that expresses what's going on much better (at which point downstream
attribute logic will become much easier to cross reference against
the class field analysis code).

One of the problems is that I need more of the logic to be match-based,
ideally with as few stages of processing as possible: this is how we can
convert the code into an enum-to-enum form that preserves information, as
opposed to the current logic that pulls apart the enum and then mutates
and/or redefines a bunch of free-floating variables.

With that big picutre in mind, I want to consolidate the first two big
statements (the one on field_definition and the one defining value_ty)
into one.

But to do that and have good readability, it would be much easier to
'flatten' the divergent logic for handling when there is or is not
a direct annotation type into the match. This can't be done using just
a match, because we have to `get_idx` on the annotation idx in order
to find out whether it has a type (it might be a qualifier-only
annotation like a bare `ClassVar`). As a result, we need to do it
in a separate block.

The extra match is a bit ugly, but I'll clean it up in a stacked diff
by moving it to a helper; it's worth noting that the "primary" match
is now quite a bit easier to skim: the duplicated annotation handling,
which was forced when we used a flat match but can be deduped if we
use a double-match, was actually most of what made the "main" match
defining `value` hard to read.

Reviewed By: grievejia

Differential Revision: D87559430

fbshipit-source-id: e533fbba899b413d30b5d2d62cc5b0b482761af1