Push mangled private attr check down
Summary:
A mangled private attribute is never inherited, and it makes sense
to push that logic down.

We wind up needing to duplicate the mangled name check twice so that
we still mark annotated attributes as definitely not inherited, but
reducing the outer if/else chain to just 2 cases seems better - I
think it's slightly more readable as-is, and it also will allow me
to perform more refactors.

Reviewed By: grievejia

Differential Revision: D87559434

fbshipit-source-id: 9261f9101a583c06e8fbf456d956198ea028527f