add module name for base settings
Summary:
We want to recognize base settings as a new variant.

Note: we can't have rootmodels + base settings as parents to the same model.

runtime message is:

"`RootModel` does not support setting `model_config['extra']`"

We should look into and implement what that should translate to in pyrefly but the prority now is to eliminate the false positives.

I think the above shows that adding a new variant is a sound thing to do.

https://github.com/facebook/pyrefly/issues/1639

Reviewed By: rchen152

Differential Revision: D87591010

fbshipit-source-id: f62e17c47f05d8dd02b8504cff86b5c46f20371f