Use ClientRequestHandle to simplify request-response pairs
Summary:
When the test client makes a request, we generally always wait for the expected response. This diff introduces a fluent API for this allowing us to call `send_request(..).expect_response(..)`.

Because the request handle keeps track of the request type, we can also avoid needing to provide the type again to await the response.

The request handle also keeps track of the request ID, meaning we can eliminate explicit request IDs in many cases.

Reviewed By: yangdanny97

Differential Revision: D87611404

fbshipit-source-id: 9d550fb51c31ac22a6a4c663644931a6c058df49