preserve parameter requiredness in Type::Concatenate
Summary:
Because the user-written `Concatenate[]` type only allows specifying types, the prepended params from user-written Concatenate are always required.

This also affects to our internal representation of param specs. When we generate parameter lists w/ param specs, we sometimes synthesize a `Type::Concatenate`. In those cases, the prefix is always treated as required, even if it's optional in the function declaration.

The fix is to update our representation of Type::Concatenate to consider requiredness, though this information is not exposed to the user.

fixes https://github.com/facebook/pyrefly/issues/1469

Reviewed By: stroxler

Differential Revision: D87661587

fbshipit-source-id: 9cd24d6d671af23082256ecf2523c61b2c0aaf26